moved default state value and helpers in out of filter component
