+++ src/utils/globals.tsx
@@ -1,5 +1,6 @@
 import { Tour } from "../models/Tour";
 import { i18n } from "i18next";
+import { Dayjs } from "dayjs";
 
 export function getTourLink(
   tour: Tour,
@@ -190,4 +191,10 @@ export const areSetsEqual = (set1: Set<any>, set2: Set<any>): boolean =>
 
 export function hasContent(obj: unknown): boolean {
   return obj != null && Object.keys(obj).length > 0;
+}
+
+export function getDurationAsHours(dayJsObject: Dayjs | null) {
+  if (dayJsObject) {
+    return dayJsObject.hour() + dayJsObject.minute() / 60;
+  }
 }
